+++ servicecatalog_puppet/waluigi/constants.py
@@ -1,0 +1,6 @@
+COMPLETED = "COMPLETED"
+NOT_SET = "NOT_SET"
+ERRORED = "ERRORED"
+QUEUE_STATUS = "QUEUE_STATUS"
+RESOURCES_REQUIRED = "resources_required"
+CONTROL_EVENT__COMPLETE = "CONTROL_EVENT__COMPLETE"
