+++ servicecatalog_puppet/waluigi/dag_utils.py
@@ -1,0 +1,52 @@
+import logging
+
+import networkx as nx
+
+from servicecatalog_puppet import constants
+from servicecatalog_puppet.waluigi.constants import (
+    COMPLETED,
+    NOT_SET,
+    ERRORED,
+    QUEUE_STATUS,
+)
+
+logger = logging.getLogger(constants.PUPPET_SCHEDULER_LOGGER_NAME)
+
+
+def build_the_dag(tasks_to_run: dict):
+    g = nx.DiGraph()
+    print("-- BUILDING THE DAG!!!")
+    for uid, task in tasks_to_run.items():
+        g.add_nodes_from(
+            [(uid, task),]
+        )
+        for duid in task.get("dependencies_by_reference", []):
+            if tasks_to_run.get(duid):
+                g.add_edge(uid, duid)
+            else:
+                logger.debug(
+                    f"{duid} is not in the task reference - this is fine when running in spoke execution mode and when the task was executed within the hub"
+                )
+
+    for uid, task in tasks_to_run.items():
+        if task.get(QUEUE_STATUS, NOT_SET) == COMPLETED:
+            try:
+                g.remove_node(uid)
+            except nx.exception.NetworkXError as e:
+                pass
+
+        elif task.get(QUEUE_STATUS, NOT_SET) == ERRORED:
+            print(
+                f"looking at task {uid} with status {task.get(QUEUE_STATUS, NOT_SET)}"
+            )
+            for n in nx.ancestors(g, uid):
+                try:
+                    g.remove_node(n)
+                except nx.exception.NetworkXError as e:
+                    pass
+            try:
+                g.remove_node(uid)
+            except nx.exception.NetworkXError as e:
+                pass
+
+    return g
